requirements.txt
discord.py
requests
PyMuPDF
google-cloud-storage
python-dotenv
pytz
matplotlib
firebase_admin
google-cloud-secret-manager